pratica02/automatos-python/requirements.txt
automata-lib==9.0.0
cached_method==0.1.0
frozendict==2.4.6
networkx==3.4.2
typing_extensions==4.13.1
